moved header to App.jsx file
