+++ src/Footer_Component/Footer.css
@@ -1,7 +1,6 @@
 .footer {
     background-color: white;
     padding: 40px 0;
-    position: fixed;
     bottom: 0;
     width: 100%;
     box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.1);
